QB15-010 Enable test as it is not crashing
The test I am enabling was crashing due to a spurious missing dependency.
Both the spurious computation and the crash were linked with this TN but
the test is not failing anymore and the spurious missing dependency is now
gone. I am therefore enabling the test and updating its output.

* testsuite/gnatprove/tests/O414-015__flow_constants_in_contracts
Enable test and update its output. There are new checks due to the work
done under R321-005. Due spurious missing dependency on const4.ads:19:23 is
now gone.

Change-Id: I766b3f7c3a6ce43c611fa724c12e6fe7972a8710