RA02-067 restore robust behaviour of Get_Array_Index
Since a recent modification, the Get_Array_Index routine assumed that
assignments to an array object within a loop can only happen via an indexed
component. This is not true, because an array can be also assigned as a
whole, e.g.:

   A := (others => 123);

or as a slice, e.g.:

   A (1 .. 3) := (others => 123);

* gnat2why/flow/flow-control_flow_graph.adb
(Get_Array_Index): relax contract and restore previous behaviour.

* testsuite/gnatprove/tests/RA02-067__flow_array_init/
Enable.

Change-Id: Id9435942287100d667394316b58a6757788a5fba