RA01-034 Ignore Compile_Time_Error/Warning pragmas
GNATprove does not have sometimes the precise information of the compiler
about size of types and objects, so that it cannot evaluate the expressions
in pragma Compile_Time_Error/Warning the same way the compiler does.
Thus, these pragmas should be ignored by GNATprove, as it can neither
verify them nor assume them (if the expression cannot be evaluated at
compile time, then the semantics for GNAT is to ignore them).

The change is implemented in the frontend, here we simply update the code
dealing with pragmas, the doc and tests.

* docs/lrm/source/language-defined-pragmas.rst
Document that pragmas are supported but ignored.

* gnat2why/flow/flow-control_flow_graph.adb
* gnat2why/spark/spark_definition.adb
* gnat2why/why/gnat2why-expr.adb
Raise a Program_Error if such pragmas are encountered, as they should be
rewritten to null statements by the frontend in GNATprove mode.

* testsuite/gnatprove/tests/O429-042__stdlib_tasking/test.out
Update test output.

Change-Id: I22aa1bf0362b124f9aa6c3657bd5afc99c4ada1f