RA02-067 Avoid crash related to array slices
This patch only allows to assign to an array thorugh a procedure/entry call
having an array slice as an actual parameter. This avoids crashing on
relevant tests but does not yet add the support for those.

* gnat2why/flow/flow-control_flow_graphs.adb
(Fully_Defined_In_Original_Loop): we now do not crash when we try to assign
an array through array slices.

* testsuite/gnatprove/tests/P915-005__flow_array_initialization_via_entry
* testsuite/gnatprove/tests/P915-005__flow_array_initialization_via_proc
Enable tests. No longer crash. Output is spurious as we are still not
supporting assignment through slices.

Change-Id: I255ea8eaf1a603c810ac9f72c9e95d5e3d8e7c08