RA04-011 Support attribute Object_Size
Attribute Object_Size can be used with GNAT to specify or query the
default size of objects of a given subtype, unless the Size attribute
is specified on the object itself. This was previously suppported by
querying X'Size for an object X of a subtype S, but S'Object_Size was
not supported. Now support it too.

This is also the occasion for cleaning up the support for other
representation attributes Size, Value_Size, Component_Size and Alignment.
While some of these can be queried for objects, the value of the object
plays no role in deciding the value of the attribute, only its subtype
and possibly explicit attributes on the variable play a role.

* gnat2why/spark/spark_definition.adb
(Mark_Attribute_Reference): Attributes Object_Size and Value_Size are
 now supported.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Deal in the same case with attributes Size, Value_Size
 and Object_Size. Fix support for Component_Size and Alignment.

* gnat2why/why/why-atree-modules.adb
* gnat2why/why/why-gen-expr.adb
* gnat2why/why/why-gen-names.adb
* gnat2why/why/why-gen-names.ads
(Insert_Type_Symbols): Adapt to new functions.

* gnat2why/why/why-gen-records.adb
(Declare_Attributes): Remove need for a type parameter.

* share/spark/theories/ada__model.mlw
(Array theories): Remove functions with prefix value/object for
 component_size and alignment. Keep only one version for the type.

* testsuite/gnatprove/tests/P511-009__test_attributes/test.out
* testsuite/gnatprove/tests/OA20-032__refinement/test.out
Update output.

Change-Id: I12b08ef8bb825e9734af3ea8ef51caa44a512418