Entity_Name point to unique constants
(no-tn-check)

When flow analysis converts Entity_Name to Entity_Id (wrapped in a Flow_Id),
it calls Find_Entity followed by Unique_Entity to get the canonical view of
a constant. For consistency, we should also get the canonical view in other
calls to Find_Entity, e.g. in Is_Opaque_For_Proof.

However, it seems simpler to call the Unique_Entity once, when the mapping
from Entity_Id to Entity_Name is registered.

This doesn't seem to affect the testsuite, but the semantics could be subtly
affected when Has_Variable_Input_Internal calls Entity_In_SPARK with a
different view of the same constant. Even if this happens, this patch
implements what was always intended.

Also, it no longer registers quantified loop parameters (e.g. objects in
expressions like "(for all Elem of Set => ...)", because they can't appear
in the flow contracts.

* gnat2why/spark/spark_register.adb
(Process_Node): register Unique_Entity of constants and ignore quantified
 loop parameters.

Change-Id: I726f2b7db81edb2c7fc7f5d01c55e7d7db4a3e30