RA07-004 do not generate compatibility axioms for volatile functions
We generate axioms for compatibility of dispatching primitives of tagged
types. These axioms refer to the logic function declared for these
primitives if they are functions. For volatile functions, there is no such
logic function. So not generate compatibility axioms for them.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Dispatch_Compatibility_Axioms): Do not generate axioms for
 volatile functions.

* tests/RA07-004__volatile_dispatch
File test.

Change-Id: I438d92df74870fe337db74af6fc5931132c67565