RA05-042 restore marking of constants in the Global contracts
Since recently we ignored the Global contract in marking, because entities
only known to proof from that contract and not referenced explicitly in any
expression can be handled as blobs, which is more efficient. However, we
missed the subtle interaction between marking and Has_Variable_Inputs.

This patch restores marking of the Global contract, or rather just the
constants objects referenced in the Global contract, which is enough to
restore the subtle interaction. It also fixes the previous approach by
marking the Depends contract when it could be used to synthesize the Global.

* gnat2why/spark/spark_definition.adb
(Mark_Subprogram_Specification): mark constants in the Global/Depends
 contract.

Change-Id: I1388191ec299847667c58144a31a5d7e51eed97d