QA11-077 add another test to demonstrate the problem
The original test for this TN included in the opening message got magically
fixed. This test demonstrates that the underlying problem still exists.

I believe that the output for call.adb is acceptable, but for call1.adb
I would expect a message about the first call to Proc1 being ineffective.

* testsuite/gnatprove/tests/QA11-077__flow_discr/
Add; the output for call1.adb is not as expected.

* testsuite/gnatprove/tests/QA11-077__flow_discr_name/
Add; similar to the above one but with global known by an Entity_Name.

Change-Id: I38febb26ca49d23e0b63a777feb3f3892b72f693