R417-042 Update rules for Proof_In constants
Proof_In constants do not participate in data- and information-flows,
and as such do not belong to Global and Depends contracts. This is
particularly important for constants with variable inputs that could be
referred to from subtype predicates, so that the exact location of
predicate checks can be ignored in flow analysis.

* docs/lrm/source/subprograms.rst
Update verification rules regarding Global.

Change-Id: I2606668e1e25e417cba615600584313dc5221efb