QA11-077 extend test to make the implicit abstract state visible
By adding a dummy generic unit to package P and instantiating it in Main2
the AST of P's body, including the Entity_Id for the X and Y objects, is
accessible when analysing Main2. Luckily, everything works as expected,
because Is_Implicit_State relies on the Flow_Scope in all its supporting
routines.

* testsuite/gnatprove/tests/QA11-077__flow_discr_name/
Extend; output is still as expected. Added to improve the coverage.

Change-Id: If3f2e22411d0ee3ec313f9d6be6023c33722bca3