RA10-010 replace low-level checks with Defining_Entity
Apparently the low-level query that meant to give us an entity of the
main-like subprogram didn't work on N_Subprogram_Declaration of a child
subprogram; now it is replaced with Defining_Entity, which is designed to
work reliably on corner-case ASTs.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Detect_Main_Subprogram): rely on Defining_Entity (needed to fix the crash)
 and Unique_Defining_Entity (for consistency).

* testsuite/gnatprove/tests/RA10-010__flow_main_child_spec/
Enable; no output (as expected).

Change-Id: Iab4a794ceb6fb3020c4c100101918706f4e8418a