RA12-018 assume predicate in loop more often
Currently, we do not assume in a loop that the predicate of local objects
hold as part of the implicit invariant because we are not sure whether
they are initialized or not. We assume the predicate in two cases:
If a type with predicates is at least partially default initialized, the
predicate should be checked at default initialization of objects.
Assume the predicate even when objects are not known to be initialized.
Assume in loops that objects initialized at declaration are initialized.

* gnat2why/why/gnat2why-expr-loops-inv.adb
(Generate_Frame_Condition): Assume initialization at objects initialized
 at declaration.

* gnat2why/why/gnat2why-expr.adb
(Assume_Dynamic_Invariant_For_Variables): Assume the predicate if the type
 defines at least partial default initialization even when objects are not
 known to be initialized.

* gnat2why/why/gnat2why-util.adb
(Is_Initialized_At_Decl): Returns True on objects initialized at
 declaration.
(Is_Initialized_In_Scope): Renaming of Is_Initialized.

* tests/Q930-002__itp/test.out
Assertion proved due to implicit invariant.

* tests/RA12-018__init_loop
New test.

Change-Id: I89e8cfc14d8c7d8c951a9006d05456b6a67515ba