RA17-008, RA18-014 Fix and improve proof of RTE in loop invariants
Fix an unsoundness in proving RTE for loop invariants, which was
introduced when adding implicit type-based loop invariants for objects
modified in the loop. Indeed, these additional loop invariants made
it incorrect to check RTE *inside* the Why3 loop, as it was done
previously.

The resulting change in the generation of Why3 code for a loop also
improves the proof absence of RTE in the loop invariant. Instead of
inserting these checks at the beginning of the Why3 loop, duplicate the
checks before entering the Why3 loop and at the end of the Why3 loop.

In order to keep the same precision regarding the checking of RTE for
a given loop invariant in the context of the previous loop invariants,
we modify slightly the generation of the program for checking RTE of
loop invariants to insert relevant guards.

* gnat2why/why/gnat2why-expr-loops.adb
(Wrap_Loop): Change generation for loop to duplicate AoRTE check for
 loop invariant.

* gnat2why/why/gnat2why-expr.adb
(Transform_Pragma_Check): Return the program expression for the check,
 so that it can be used as guard for checking following loop invariants
 in Transform_Loop_Statement.

* testsuite/gnatprove/tests/LB27-011__overflow_other_switch/test.out
Fix unsoundness related to previous translation scheme.

* testsuite/gnatprove/tests/RA17-008__loop_invariant_checks/repro.adb
* testsuite/gnatprove/tests/RA17-008__loop_invariant_checks/test.out
* testsuite/gnatprove/tests/RA17-008__loop_invariant_checks/test.py
New test.

* testsuite/gnatprove/tests/K223-023__search/test.out
* testsuite/gnatprove/tests/L831-037__slice/test.out
* testsuite/gnatprove/tests/LB28-004__loop_variant/test.out
* testsuite/gnatprove/tests/N401-026__variant/test.out
* testsuite/gnatprove/tests/NA29-017__prove_all/test.out
* testsuite/gnatprove/tests/O512-022__fibonacci/fib.adb
* testsuite/gnatprove/tests/O722-006__functional_lists/test.out
* testsuite/gnatprove/tests/QA31-008__von_neumann_sqrt/test.out
* testsuite/gnatprove/tests/QC08-002__loopinv/loopinv.adb
* testsuite/gnatprove/tests/QC08-002__loopinv/test.out
* testsuite/gnatprove/tests/R125-020__comparison/test.out
* testsuite/gnatprove/tests/R419-026__swap_ranges/test.out
* testsuite/gnatprove/tests/R516-050__functional_lists/test.out
Improved results.

* testsuite/gnatprove/tests/M809-005__float_basics/test.out
* testsuite/gnatprove/tests/O512-022__number_theory/test.out
* testsuite/gnatprove/tests/O512-022__prefix_sum/test.out
* testsuite/gnatprove/tests/P508-001__bits_manipulation/test.out
* testsuite/gnatprove/tests/QB10-026__counterexample/test.out
* testsuite/gnatprove/tests/prime_numbers/test.out
Slightly less precise results, due to larger context for proof.

* testsuite/gnatprove/tests/*/test.out
Update outputs due to number of VCs changing.

Change-Id: I1c7fd5498bebb8e6cc5d2d1e6099f0583f4f7547