RA02-014 don't cache crashes of gnatwhy3
We don't want to cache crashes of gnatwhy3; also we know that gnatwhy3
returns 0 in normal execution. Other processes may return non-zero exit
code and we still want to cache them.

So we check for the name of the executable and the exit code before
caching the result.

* spark_memcached_wrapper.adb
(SPARK_Memcached_Wrapper): don't cache calls to gnatwhy3 with a non-zero
exit code.

Change-Id: I3671718bd032843ec77e1d65e0a6845a2b48d278