RA29-007 do not support local borrowers and traversal functions
To allow modifying a pointer through a borrower, we need some special
support that is not availlable for now. In the mean time, disallow
declaration of locala borrowers and traversal functions returning a
access-to-variable type.

* gnat2why/spark/spark_definition.adb
(Mark_Object_Entity): Local borrowers of access to variable types are not
 supported yet.
(Mark_Function_Specification): Traversal functions returning access to
 variable types are not supported yet.

tests/RA29-007__mutable_borrowers
Add test. For now traversal functions are rejected by the borrow checker.

Change-Id: I0177ca23d6546fb99d1e4dbf525175693b66f6ae