RA29-039 fix crash on for loop on record component
When the container on which iteration is done is a variable, we cannot
introduce a temporary variable for it in Why, as it may change during
the iteration. When the container is used in the logic domain, we should
use an expression from the EW_Term domain for it.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Translate the container in the term domain
when using it in logic.

Change-Id: I537c922a1c8898864d7aba05b22de2b0fead9f88
(cherry picked from commit 87b51afe4e12935089fc75ea2313f34ed427f584)