RB05-005 add test for unsoundness that is already fixed
The current output appears correct, but we got a report that some older
version of GNATprove didn't emit two checks, that are rightly unproved:

 main.adb:23:04: medium: precondition might fail
 main.adb:25:35: medium: precondition might fail

Also, release 18.2 is crashing on this code and gnat-community-2018 is
excessively slow (if it ever terminates at all, I didn't wait long enough to
know).

* testsuite/gnatprove/tests/RB05-005__unsound/
Add; output seems as expected.

Change-Id: I63d2a7c326e1df4de07c10f890ca48966f0befe2