RB09-010 print constrained attribute and discriminants of split records
When a record variable is in split form, we should mark all of its
components tp be used in counterexamples.

* gnat2why/why/gnat2why-ce_utils.adb
(Has_Prefix): New function to decide if a string is a prefix of another
 string.

* gnat2why/why/gnat2why-counter_examples.adb
(Refine_Attribute): Refine boolean values.
(Handle_CNT_Element): Use Has_Prefix.

* gnat2why/why/gnat2why-decls.adb
(Translate_Variable): Add labels to discriminant and 'Constrained
 parts of record variables.

* tests/*/test.out
Counterexample values for discriminants and 'Constrained. Note that as
these values are (often) constants, they are sometimes expressed on the
old value instead of the current one.

Change-Id: Icae94f09ba51222b432e6401706bc1e7d83cd56e