O720-013 fix naming of a-something units
GNAT considers units whose names start with "a-" as internal (because of the
DOC 8.3 file naming convention, e.g. a-textio.ads) and emit warnings when
compiling such files. In GNATprove mode those warnings are not emitted and
the GG algorithm is silently derailed.

This meant to be fixed on this TN, but got killed as minor and internal, so
for now just rename problematic files in unrelated tests. This kills
spurious warnings of the form:

  a.adb:15:44: warning: no Global contract available for "F_6"
  a.adb:15:44: warning: assuming "F_6" has no effect on global items

* testsuite/gnatprove/tests/P118-034__flow_gg_children/
* testsuite/gnatprove/tests/P118-034__flow_gg_private/
* testsuite/gnatprove/tests/R322-036__ali/main_src/
* testsuite/gnatprove/tests/R413-012__mainonly/main_src/
Rename a-... (dash) to a~... (tilde). Update outputs.

Change-Id: I77ac077503ed6d0c841cb8ce7283c76f4eb57378