only serialize "local_var" for packages
The "local variables" recorded between phase 1 and 2 is only used to
generate the Initializes contract, so it is only meaningful for packages.

Serializing it for all entities is a relic of the old GG, which kind of
special-cased local objects, but this is no longer used.

(no-tn-check)

* gnat2why/flow/flow_generated_globals-phase_1.adb
(GG_Register_Global_Info): only write "local_var" for packages.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Serialize): only read "local_var" for packages.

Change-Id: Ib7577a3ad02ef4d7d60b630d2d6f0752fc7256bf