R619-008 Use Sequence instead of New_Statement_Sequence
This tries to optimize Sequence so that it generates less succession
of "();()" in .mlw files.
Sequence now uses an Is_Void condition which returns true for void or
sequence of void: it is basically the same as Is_Essentially_Void but
it keeps the comments.

* gnat2why/why/why-gen-expr.adb
(Is_Void): Statement is void or a sequence of void
(Is_Void_List): Same as Is_Void for List.

* gnat2why/why/why-gen-progs.adb
(Sequence): Now uses Is_Void
(Sequence): Now the array is refined so that we remove the void
statements
(Sequence_Append): Uses Is_Void instead of equality.

* gnat2why/why/gnat2why-expr.adb
(Transform_Statement_Or_Declaration_In_List): Uses Sequence instead
of New_Statement_Sequence

Change-Id: Ib1640030da938bcb57fbf8f68c1f59f928c883fd