R522-032 add (a non-minimized) reproducer
The TN is about global generation in complex code hierarchy with child and
generic units, so it is best to have the entire code as a test. However, to
make things easier, our code is (almost) a flat view of two repositories and
is already preconfigured for one of the supported target architectures.

The README file explains where this code came from.

* testsuite/gnatprove/tests/R522-032__flow_libgfxinit/
Add; the "XXX must be listed in the global aspect of YYY" are spurious.

Change-Id: I83166e6ce8e52356bfcf10cb81b648c71edbc9ef