RC03-030 Add test for why3config
This adds a test that tries to create a why3.conf then add a prover to the
conf file to try to run it with gnatprove.
It then removes the executable for this prover and check that a new error
should be produced.

* testsuite/gnatprove/tests/RC03-030__conf/test.adb
(Test): Random failing test with an increment function

* testsuite/gnatprove/tests/RC03-030__conf/test.py
Added successive commands of why3config and prove for this test

* testsuite/gnatprove/tests/RC03-030__conf/toto.conf
Initially empty conf file

* testsuite/gnatprove/tests/RC03-030__conf/toto.sh
Script made to look like "z3 --version" in order to be recognized by Why3
as being of Z3 type.

Change-Id: I26849615e6b5a905898a1027e28796a64f73e303