R703-025 fix crash of spark_report
Spark_report depended on Configuration, but this was incorrect.
Configuration was only intended to be used from Gnatprove. A previous
commit added some elaboration code to Configuration, which was therefore
also executed in the spark_report binary, which was not intended and
failed.

This commit results in some slight code duplication, but we can fix this
at a later stage.

* spark_report.adb
remove dependency on Configuration unit

Change-Id: I3c82b2cc6ba31a7081f773139a97bf2d5311f85c