RB30-003 Extend explanations for output parameters
When a check is unproved, it may be due to the absence of information
in preconditions on array bounds or record discriminants from output
parameters. Explanations now take these into account.

Also fix the traversal for finding a possible explanation node, which
was stopping in some cases where it is possible to continue in fact,
like on case-statements. This leads to more explanations inside
case-statements.

Also fix the traversal from a check inside contracts attached to a
subprogram declaration, so that the search stops at the matching
subprogram declaration.

Also avoid justifying a postcondition by some missing information in
the precondition, or a run-time check inside the postcondition by
missing information in the precondition if this run-time check
involves the result of the function or some other post-state.

* gnat2why/flow/flow_error_messages.adb
(Is_Stmt_Or_Prag_Or_Decl): Include N_Later_Decl_Item as declarations.
(Get_Previous_Explain_Node): Fix function to properly find a previous
 node in various cases which previously returned Empty, like when N
 is inside a case-statement.
(Get_Explanation): Treat specially the case of output parameters of
 unconstrained type. Also handle assignment nodes in input.
 Also handle pragmas attached to subprogram declarations.

* testsuite/gnatprove/tests/*
More explanations, due both to the new handling of assignments, and
the fix in Get_Previous_Explain_Node.

Change-Id: Iff5aaed55a6bf31129682906806800bba2075686