RC12-029 assume array bounds on default initialization
Without pointers, we never needed to assume the value of the bounds of
arrays in default initialization. Indeed, the bounds were always available
through the type of the default initialized variable. For allocators, the
type used for the default initialization may differ from the expected type
so we need to assume them explicitly.

* gnat2why/why/why-gen-arrays.adb
(New_Bounds_Equality): New version which queries bounds from a type.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Init): Assume bounds of arrays.

* tests/RC12-029__pointer_array_sutype
new test

* tests/LB28-004__loop_variant/test.out
Change in counterexamples

Change-Id: I1bc0e1985510bed4f90808fc4a2944d1f0d524a9