S115-027 increase visibility for private part of generic instance
The missing link was actually present in the original design document, but
doesn't appear in its implementation. We didn't detect this before, because
this link only matters if the generic itself is a child package, which is
not a common pattern.

* gnat2why/flow/flow_visibility.adb
* gnat2why/flow/flow_generated_globals-phase_2-visibility
(Connect): add missing link (for both phase 1 and 2).

* testsuite/gnatprove/tests/S115-027__global_gen/
Enabled test; no output, as expected.

Change-Id: Ie573cfc695b834ffba3cd1a7597100ef94163c59