R329-039, RC14-043 Support SPARK_Mode Off inside subprograms
Historically, SPARK_Mode was only allowed at library-level, so that parts
with SPARK_Mode Off can be tested more easily. This turns out to be
restrictive for some uses, so this patch allows the use of SPARK_Mode Off
to exempt parts of a subprogram from analysis. This is particularly
useful for instantiating generic packages with SPARK_Mode Off bodies inside
a subprogram.

SPARK_Mode On cannot be used inside subprograms, except for the special case
of SPARK_Mode in instantiations, in which case it has no effect (either the
context already provides SPARK_Mode On and the entity is analyzed, or it
does not and the entity is not analyzed).

* docs/ug/en/appendix/implementation_defined_pragmas.rst
Modify the SPARK_Mode definition rules.

* testsuite/gnatprove/tests/R329-039__spark_mode_off/gen.adb
* testsuite/gnatprove/tests/R329-039__spark_mode_off/gen.ads
* testsuite/gnatprove/tests/R329-039__spark_mode_off/inst.adb
* testsuite/gnatprove/tests/R329-039__spark_mode_off/p.adb
* testsuite/gnatprove/tests/R329-039__spark_mode_off/test.out
* testsuite/gnatprove/tests/R329-039__spark_mode_off/test.py
* testsuite/gnatprove/tests/R329-039__spark_mode_on/p.adb
* testsuite/gnatprove/tests/R329-039__spark_mode_on/test.out
* testsuite/gnatprove/tests/R329-039__spark_mode_on/test.py
New tests.

Change-Id: I8915f38b9f629a2d11dadbc151953508cf2014a0