S107-014 remove explicit USE clauses for global effects
When translating subprogram specs we were adding an explicit USE clause for
each of its globals, but for functions we only used names of the globals in
binders and didn't refer to their why3 objects, e.g. for an Ada routine:

   function Has_No_Piece return Boolean with Global => Engine.State;

we had:

   module Engine__has_no_piece
    use        Engine__state                   (* <<< unnecessary *)
    use import "_gnatprove_standard".Main
    use        "_gnatprove_standard".Main

    function has_no_piece
      (engine__state : Main.__private) : bool
    (* ^^^^^^^^^^^^^   <<<<< this is just a name, not a reference *)
   end

This was a trouble when changing the expansion of abstract states, because
in code for deciding order between modules we failed to recognize
dependencies between modules for a function and for its globals.

For procedures those USE clauses are added automatically by Close_Theory,
where we first Record_Dependencies and then Add_Definition_Imports. For
procedures Record_Dependencies actually records something, because there the
globals are actually referenced, while for functions they only appear in
binders.

* gnat2why/why/gnat2why-subprograms.adb
(Translate_Subprogram_Spec): don't add dependencies for effects.
(Translate_Expression_Function_Body): same as above.

(Add_Dependencies_For_Effects): remove.

* gnat2why/why/why-inter.ads
* gnat2why/why/why-inter.adb
(Add_Effect_Import): remove.

Change-Id: I77c9ca74b20bea43d42074b38b0d851977893620