simplify Is_Relevant_Node_For_Imports
Routine Is_Relevant_Node_For_Imports was always called with the default
value of its Filter_Entity parameter (Empty). Also, all of its tests when N
was an entity were always True. Finally, the _Node_ infix in its name was
repeating the type of its parameter N, which itself was not precise (because
it could denote either a node or an entity).

Now all of these always true conditions are executed as ghost code and the
body is as simple as it could.

(no-tn-check)

* gnat2why/why/why-inter.adb
(Is_Relevant_Node_For_Imports): simplified.

Change-Id: I169ea85b04822597f9f5b53fa1499e2e860fa42f