R216-040 respect visibility of Refined_Post
Routine Has_Visibility_On_Refined was effectively checking whether the spec
of a subprogram is visible at the point of call; this is always the case
(because subprogram can be called only if it is visible). Now it checks
whether the body of a subprogram is visible (or a place where such a body
should really be, e.g. when there is no body yet, or it is an expression
function declared in a package spec).

* gnat2why/why/gnat2why-expr.adb
(Has_Visibility_On_Refined): add contract to spec; fix body.

* testsuite/gnatprove/tests/R216-040__peek_on_refined_post/test.out
Enable; add output.

* testsuite/gnatprove/tests/MA29-016__flow_refinement/test.out
Update output; Refined_Post of Test_13, which is now proved, is exactly the
same as Post of Test_11, which is the only procedure called by Test_13.

Change-Id: Idde6e907151a48cd22ebd0491e6d4e20104ca977