S201-050 Fix explanation inside generated procedures
The search for an explanation should not apply inside a generated procedure
such as a DIC procedure or an invariant procedure.

Querying for Global contract on DIC (and Type_Invariant) procedures is
semantically ill-defined. This patch clarifies that such queries are not
expected and the DIC/Type_Invariant procedures are not explicitly analysed
by flow.

* gnat2why/flow/flow_error_messages.adb
(Get_Explanation): Skip explanation inside DIC or invariant procedure.

* gnat2why/flow/flow.ads
(Flow_Analyse_Entity): clarify that DIC and Type_Invariant procedures are
 not expected.

* gnat2why/flow/flow_utility.ads
(Get_Globals): same as above.

Change-Id: I8973d17f587bb7c2260871e14b559164daca77c4