R202-034 fix package in protected type
When an entity in a protected object/type is processed, need to
set its containing protected type. This was not done for
packages, now done.

* spark_util-subprograms.ads
(Containing_Protected_Type): can also be called for packages
* gnat2why-expr.adb
(Assume_Declaration_Of_Entity): add assertion that would have simplified
  analysis
* gnat2why-subprograms.adb
(Generate_VCs_For_Package_Elaboration): if the package is in a
  protected_type, need to set containing protected type

* R202-034__pkg_in_prot_op:
  proof enabled in test

Change-Id: Iec19ec04ccc84365f372b6db17a9fb47cf23cc64