R502-058 readapt tests for corrupted ALI files
* testsuite/gnatprove/tests/P307-006__tool_bad_cleanup/
Preserve ALI timestamp, as otherwise it is recreated when rerunning the tool
and the entire test becomes useless. Change test discriminant to packaged,
because the gprbuild decides when to rerun phase1 based on a tool version
tag that only appears in wavefront & releases, not in developer builds.

* testsuite/gnatprove/tests/O720-004__corrupted_ali_file/
Preserve ALI timestamp; remove unreferenced "target_ali2" object.

Change-Id: Ia06788f0fdd5acd06b67107fb5303d610cdf18fe