S213-020 add test for child package and Initializes
In the original reproducer there was Repro.State, Repro.B.State and
Repro.C.State; in this test the latter two are renamed to BState and CState,
just to make this easier to follow.

Also, type Byte is a simple integer and not Interfaces.Unsigned_8, to not
pull the Interfaces into visibility graphs, which makes them easier to read.

* testsuite/gnatprove/tests/S213-020__flow_child_init/
Add; spurious high check from the flow analysis.

Change-Id: Id477a77f19930b16953d64d92263b1d277a07ff9