S213-020 down-project RHS of the Initializes to the current scope
The RHS of an Initializes contract of a package might reference an abstract
state whose constituents are visible in that package body. Such abstract
states must be down-projected, just as it happens for the Global contract.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): down-project items from the RHS of an Initializes contract.

* testsuite/gnatprove/tests/S213-020__flow_child_init/
Spurious warning is gone; add pragma Elaborate as required.

Change-Id: I136e6760039187d6447feeb0933fa5f7de07865c