S117-042 do not follow "barrier" edges in DFS for dead code
When traversing the CFG to check which vertices are dead we shall not follow
the "barrier" edges, i.e. paths corresponding to executions of entry bodies
with barrier expression evaluated to False and to executions of CASE
statements with a single alternative where this alternative is not selected.

* gnat2why/flow/flow-analysis.adb
(Find_Dead_Code): skip children of the auxiliary "barrier" vertex.

* testsuite/gnatprove/tests/S117-042__flow_case/
Add a minimized version of ACATS where "this statement is never reached"
warnings got killed and update the output accordingly.

Change-Id: I75da3d0a602c9b9afebf4f3cc361db0cc6eeb1b7