S220-009 Use the object directory for prepare_task
Use the object directory for the prepare_task phase of gnatwhy3.

* gnatprove/gnatprove.adb
(Prepare_Why3_Manual): Get the Obj_Dir and use it when switching for
prepare_task.

* testsuite/gnatprove/tests/S220-009_check_dirs/*
Add test which checks the number/type of directory change during Coq
with --proof-dir test. This output may change (hopefully) when we change
the change-directory logic in gnatprove.

Change-Id: I8c2710b791a15625149fc5f951f3fc90399e2ba9