Q612-030 Update outputs, firstfew tests outputs
New updating outputs of testsuite for newsystem. This includes:
- Change of prover in proofs
- Minor change in counterexamples
- minor explained removal of counterexamples
- New checks that comes after a raise Program_error (they were probably
optimised away before)
- Raising steps to solve diff sometimes

* L323-005__recursive
Minor removal. Cannot reproduce the new diff locally

* M611-010__no_return_illegal
Minor new proved check that corresponds to an inaccessible point of the
program.

* M613-039__expr_actions
One minor change in provability (one proved check). The rest is minor
change in counterexampels.

* N122-021__succ_floats/succ_floats.adb
Ease the future investigation of regressions on this file by adding
assert pass/fail

* N122-021__succ_floats/succ_floats.*
Raise output removes regressions.

* N614-001__crash/test.out
Minor Z3 instead of CVC4

* N624-020__perm_occ
Cvc4 proves instead of Z3

* N624-020__perm_rem
Minor diff in provability. 138:22 is indeed provable.

* N624-030__funcs_on_containers
Minor Z3 instead of CVC4

* N709-001__contracts
Minor altergo instead of cvc4. One more counterexample value.

* N709-001__tagged_discr
Minor Z3 instead of cvc4

* N917-051__longest_common_prefix
Re-add Z3 which was historically used. Local testing remove the diffs. More
stuff will be proved in mailserver -> these are oks.

* NB04-040__private_base
Diff in counterexample: acceptable

* NB24-034__flow_spurious_ghost_warning
Only flow is run for this test so provability diff are ok.

* NC06-004__init
This range check is indeed true. R is bounded 0..10 so each time it is
assigned, it is protected by the if just before it.

* O217-033__liskov
Cvc4 to altergo: OK

* O227-007__float_double_inconsistency
This is more of a regression of Codepeer. It works with steps raised to
15500: it should not take long because there is only one prover and one
check.
no diff when steps are raised

* O324-035__time_triggered_response
One check that replace cvc4 with altergo -> ok

* O401-034__float_pred
It looks correct from the function.

* O403-007__attr_constrained_private
Change of provers: Ok

* O407-031__flow_gg_nested_problem
Minor new check that corresponds to something that was always True.

Change-Id: I2c504b6b80316c385ed9906a602724ff9c34ad24