Q612-030 third wave of test updates
New updating outputs of testsuite for newsystem. This includes:
- Change of prover in proofs
- Minor change in counterexamples
- minor explained removal of counterexamples
- New checks that comes after a raise Program_error (they were probably
optimised away before)
- Raising steps to solve diff sometimes

* O512-022__discovery_nosegear
Ok the test is apparently to test Altergo. cvc4 proves the goals.

* O512-022__fibonacci
One less proved check. Ok, the important check in this test are marked.

* O512-022__vscomp2014_partition_functional
Worrying duplication of precondition but it already happened for 1469:14
and 1476:23.
The rest is just more proved goals.

* M809-005__float_basics
Updated sessions, some more proved tests

* OA20-032__refinement
More proved stuff and more counterexamples -> ok.

* OA21-028__failsafe
One more proved goal. Diff in number of VCs.

* P512-044__exist_on_bv
Few provability differences

* P708-032__riposte_basic_z3_CE
Minor diff in counterexample traces.

* P719-001__force
This apparently test option --force which is not in contradiction with this
diff.

* PB16-052__float_invariant
The test says undischarged VCs are not to be accepted in this test.
Raising the steps to 3000 solve the problem locally. Note that there are
already undischarged VCs in attempt_3.

* PB18-021__tools_replay_with_custom_prover
Fix the replay function in the test so that we can call update-session.
After regeneration of the session, the diff disappeared.

* PC07-014__float_conversion
Minor diff in counterexample (trivial counterexamples that should
disappear). Minor diff in provers and provability in a file that already
contains countless undischarged VC -> ok

* Q313-018__system_counterex
Minor counterexample trace change

* Q313-019__elementary_ce_float
Change in float counterexamples could be explained by lack of stability of
float counterexamples.

* Q424-012__succ_floats_copy_z3
This is a copy of another example which was fully proved (namely
N122-021__succ_floats). This file is used only for counterexamples with z3.
So, change in provability are oks.

* Q615-017__factorial
Add z3 and steps to prove what was lost

* Q915-037__counterex
Minor counterexample trace change.

* Q930-002__itp
Minor change in provers

* QA11-003__itp_server_test
Probably a forgotten update of output. Having this output means that
gnat_server is ok (Task + CST_NODE...).

* QA11-057__counterexamples
The 'Result in this file are actually left value in this case. So it is
good that they disappear.

* QA25-058__z3_riposte_arrays
Minor counterexamples diff

* QA31-008__con_neumann
Proof was broken a long time ago. We don't care about this.

* QB10-026__counterexample
Diffs on provability and counterexamples. But this is a counterexample test
so this is minor.

* QC11-016__for_of_cntex
Minor change in counterexamples. Already reported duplication of
precondition.

* R125-020__comparison
Minor diff in counterexamples.

* R508-015__ce_branch
This is the if-branching test. I checked that the diffs are ok with regards
to not having value in "dead branch".

* R511-031__counterexample
Minor diff in counterexample

* R514-028__ce
Minor diff in counterexample

* R514-028__ce
Minor diff in counterexample

* R709-026__membership
Change in prover -> ok

* R807-012
Minor change in counterexamples. Minor change in provability that can be
reproved using manual proof.

* R906-030__recursion
Minor diff in provers

* RC17-014__init_by_proof
Minor diff in provers

* prime_numbers
More proved stuff. The proved assertions should be correct as this example
is part of the "SPARK examples".

* random
Raise steps is enough to get the proof back

* riposte_*
Already investigated minor diff in counterexamples.

* string_search
Raising the steps (on my local setup) solve the Undischarged VCs problem.

* tagged_stacks
Minor prover diffs

* tetris
Minor prover diff

Change-Id: Ic2d68081b0145597991062a9289319ca3120d727