Q612-030 Update of tests
More updates:
- diffs with last mailserver,
- some diffs built locally on my machine,
- some VC not proved anymore.

* K624-029__amortize_queue
diff of prover

* M603-018__float
Minor provability diff. Raising steps would be enough to prove the whole
file -> probably not the purpose of the test

* N122-021__succ_floats
Increase steps for Gaia tests

* O227-007__float_double_inconsistency
Increase steps for Gaia tests

* O331-029__float
Apparently a test for Z3 with steps=1. Not proving much is expected.
Filing the diff as is.

* O512-022__number_theory
Minor provability differences.

* O831-003__do333_mode_logic_qgen
Minor counterexample lost

* P405-006__mod_arithemtic_lemmas
Provability is most likely not important in this test. One more VC
discharged.

* P508-001__bits_manipulation
Some new assertions proved locally. The tests does not seem to contain
smoke -> ok

* P630-037__keccak
Run locally. Some VC are now split in two checks: this is something that
has already been observed, the number of goals is the same so it should
not be too bad.
Some minor provability and counterexamples changes.

* PA25-037__bit_set
Increase steps to get the same results in GAIA as locally.

* PB16-052__float_invariant
More stuff proved and differences in provers

* QB30-008__session
The counting and names of goals changed with newsystem. This is no
surprising that this test change. The new value looks ok at first glance.

* R419-026__swap_ranges
Change in counterexamples

* R509-011__higher_order
Updated with local results. Some prover changes.

* R509-011__higher_order_proof
Updated sessions locally. Only one new regression.

* S212-028__update
Minor diff in counterexamples

* proofinuse__lat_long
One less VC proved. Accept the diffs. The test already contains unproved
checks.

* riposte_array_applications
Restore the sessions. Lost some proofs but we don't mind because this is
originally a test for counterexamples.

Change-Id: I4cb744f410a386ca17c24312f9b0a8b62273939d